Changed FinalGrade property from Enrollment and Evaluation classes to double. Finished test data for the tests.

+++ ADYC.Model/Enrollment.cs
@@ -9,7 +9,7 @@ namespace ADYC.Model
     public class Enrollment
     {
         public int Id { get; set; }
-        public Decimal? FinalGrade { get; set; }
+        public Double? FinalGrade { get; set; }
         public GradeLetter? FinalGradeLetter { get; set; }
         public string Notes { get; set; }
         public bool IsCurrentEnrollment { get; set; }
